AG-11087 - Split gallery examples from examples.spec.ts.
