+++ packages/ag-charts-website/e2e/examples.spec.ts
@@ -1,30 +1,7 @@
-import { expect, test } from './fixture';
-import { getExamples, gotoExample, setupIntrinsicAssertions, toExamplePageUrls, toGalleryPageUrls } from './util';
+import { ExampleOverrides, convertPageUrls, createTestCase } from './examples-util';
+import { test } from './fixture';
+import { getExamples, setupIntrinsicAssertions } from './util';
 
-type Status = 'ok' | '404';
-type ClickOrder = 'normal' | 'reverse';
-
-type ExampleOptions = {
-    pagePath: string;
-    url: string;
-    example: string;
-    framework: string;
-    status: Status;
-    clickOrder: ClickOrder;
-    skipCanvasUpdateCheck: boolean | string[];
-    ignoreConsoleWarnings: boolean;
-};
-
-type ExampleOverrides = {
-    frameworks?: string[];
-    status?: Status;
-    skipFrameworks?: boolean;
-    clickOrder?: ClickOrder;
-    skipCanvasUpdateCheck?: boolean | string[];
-    ignoreConsoleWarnings?: boolean;
-};
-
-const ignorePages = ['benchmarks', /.*-test/];
 const exampleOptions: Record<string, Record<string, ExampleOverrides>> = {
     gallery: {
         '*': { frameworks: ['vanilla', 'typescript'] },
@@ -126,62 +103,15 @@ const exampleOptions: Record<string, Record<string, ExampleOverrides>> = {
     },
 };
 
-function convertPageUrls(path: string) {
-    const astroPath = path.split('content/').at(1)!;
-    const [pagePath, examplePath] = astroPath.split('/_examples/');
-    const example = examplePath.replace(/\/[a-zA-Z-]+\.ts$/, '');
-
-    const page = pagePath.replace(/^docs\//, '');
-    const pages = pagePath === 'gallery' ? toGalleryPageUrls(example) : toExamplePageUrls(page, example);
-
-    if (ignorePages.some((m) => (typeof m === 'string' ? m === page : m.test(page)))) {
-        return [];
-    }
-
-    const {
-        frameworks,
-        status = 'ok',
-        clickOrder = 'normal',
-        skipCanvasUpdateCheck = false,
-        ignoreConsoleWarnings = false,
-    } = {
-        ...exampleOptions[page]?.['*'],
-        ...exampleOptions[page]?.[example],
-    };
-
-    return pages
-        .filter((r) => frameworks?.includes(r.framework) !== false)
-        .map(
-            ({ url, example: pageExample, framework }): ExampleOptions => ({
-                pagePath,
-                url,
-                example: pageExample,
-                framework,
-                status,
-                clickOrder,
-                skipCanvasUpdateCheck,
-                ignoreConsoleWarnings,
-            })
-        );
-}
-
 test.describe('examples', () => {
     const config = setupIntrinsicAssertions();
 
     const examples = getExamples();
 
     for (const { path, affected } of examples) {
-        for (const opts of convertPageUrls(path)) {
-            const {
-                url,
-                status,
-                framework,
-                pagePath,
-                example,
-                clickOrder,
-                skipCanvasUpdateCheck,
-                ignoreConsoleWarnings,
-            } = opts;
+        for (const opts of convertPageUrls(path, exampleOptions)) {
+            const { framework, pagePath, example } = opts;
+            if (pagePath === 'gallery') continue;
 
             // eslint-disable-next-line @typescript-eslint/unbound-method
             const testFn = affected ? test : test.skip;
@@ -190,54 +120,7 @@ test.describe('examples', () => {
                 test.skip(!affected, 'unaffected example');
 
                 test.describe(`Example ${pagePath}: ${example}${affected ? '' : ' (!!!SKIPPED!!!)'}`, () => {
-                    if (status === 'ok') {
-                        testFn(`should load ${url}`, async ({ page }) => {
-                            test.slow(framework === 'angular', 'allow more time for Angular load times');
-
-                            config.ignoreConsoleWarnings = ignoreConsoleWarnings;
-
-                            // Load example and wait for things to settle.
-                            await gotoExample(page, url);
-
-                            // Check we're dealing with a single canvas, otherwise things get tricky!
-                            const canvases = await page.locator('.ag-charts-wrapper').all();
-                            if (canvases.length > 1) return;
-                            const canvas = canvases[0];
-
-                            // Try pressing the buttons to see if any errors are thrown.
-                            const buttons = await page.locator('.toolbar > button').all();
-                            if (clickOrder === 'reverse') buttons.reverse();
-
-                            for (const button of buttons) {
-                                const sceneRenderCount = Number(await canvas.getAttribute('data-scene-renders'));
-
-                                await button.click();
-
-                                const skip =
-                                    skipCanvasUpdateCheck === true ||
-                                    (Array.isArray(skipCanvasUpdateCheck) &&
-                                        skipCanvasUpdateCheck.includes((await button.textContent()) ?? 'unknown'));
-                                if (skip) {
-                                    await page.waitForLoadState('networkidle');
-                                } else {
-                                    await expect
-                                        .configure({
-                                            message: `Pressing button ${await button.textContent()}`,
-                                        })
-                                        .poll(async () => Number(await canvas.getAttribute('data-scene-renders')))
-                                        .toBeGreaterThan(sceneRenderCount);
-                                }
-                            }
-                        });
-                    }
-
-                    if (status === '404') {
-                        testFn(`should 404 on ${url}`, async ({ page }) => {
-                            config.ignore404s = true;
-                            await page.goto(url);
-                            expect(await page.title()).toMatch(/Page Not Found/);
-                        });
-                    }
+                    createTestCase(testFn as any, opts, { initialScreenshot: false, ...config });
                 });
             });
         }
